GPIO compilation fix (#1049)
+ updated build.config to remove not needed module exclusion

IoT.js-DCO-1.0-Signed-off-by: Konrad Lipner k.lipner@samsung.com

+++ src/platform/tizenrt/iotjs_module_gpio-tizenrt.c
@@ -25,7 +25,7 @@ void iotjs_gpio_open_worker(uv_work_t* work_req) {
          _this->direction, _this->mode);
 
   // Open gpio pin
-  _this->gpio_context = iotbus_gpio_open(_this->pin);
+  _this->gpio_context = iotbus_gpio_open((int)_this->pin);
   if (_this->gpio_context == NULL) {
     req_data->result = false;
     return;
